Added warning protocol

+++ src/ti/val.c
@@ -1121,7 +1121,7 @@ int ti_val_to_file(ti_val_t * val, FILE * f)
     return -1;
 }
 
-void ti_val_may_change_pack_sz(ti_val_t * val, size_t * sz, uint8_t * nest)
+void ti_val_may_change_pack_sz(ti_val_t * val, size_t * sz, size_t * nest)
 {
     switch ((ti_val_enum) val->tp)
     {
